requirements.txt
gradio
groq
elevenlabs
gtts
python-dotenv
pydub
speechrecognition
ffmpy
pillow
numpy
pandas